Third attempt
Kinda complicated, found another data discrepancy where the halfEdge in
edges will be none of the half-edges pointing back to it.

Added a little fix method for that.

Now the code attaches "Shape" instances to the edges and tracks the
points as well as the edges crossed by that shape.

This way we can make sure that we update the shape instances right in
all cases.

Finally we can then draw each shape in one random color.