Improve performance
We seem to have some random cases where the performance suddenly drops
off.. Not sure if it's just a bug or a complexity issue.

I tend to believe the latter. This is also supported by this code change
improving things.

We check the shapes attached to the half edges and only add them if one
of the half edges does *not* already have a shape attached to it.